+++ src/main/java/com/jmeter/websocket/plugin/elements/samplers/WebsocketMessageSamplerGui.java
@@ -1,4 +1,4 @@
-package com.jmeter.websocket.plugin.samplers;
+package com.jmeter.websocket.plugin.elements.samplers;
 
 import org.apache.jmeter.gui.util.HorizontalPanel;
 import org.apache.jmeter.testelement.TestElement;
