+++ src/main/java/com/jmeter/websocket/plugin/elements/samplers/configurers/AbstractWebsocketSamplerConfigurer.java
@@ -1,7 +1,8 @@
-package com.jmeter.websocket.plugin.configurers;
+package com.jmeter.websocket.plugin.elements.samplers.configurers;
 
-import com.jmeter.websocket.plugin.samplers.AbstractWebsocketSampler;
-import com.jmeter.websocket.plugin.samplers.AbstractWebsocketSamplerGui;
+import com.jmeter.websocket.plugin.elements.helpers.Configurer;
+import com.jmeter.websocket.plugin.elements.samplers.AbstractWebsocketSampler;
+import com.jmeter.websocket.plugin.elements.samplers.AbstractWebsocketSamplerGui;
 
 public class AbstractWebsocketSamplerConfigurer implements Configurer<AbstractWebsocketSampler, AbstractWebsocketSamplerGui> {
