+++ src/test/groovy/com/jmeter/websocket/plugin/elements/samplers/modifiers/AbstractWebsocketSamplerModifierSpec.groovy
@@ -1,10 +1,10 @@
-package com.jmeter.websocket.plugin.modifiers
+package com.jmeter.websocket.plugin.elements.samplers.modifiers
 
 import com.jmeter.websocket.plugin.JmeterAbstractSpec
-import com.jmeter.websocket.plugin.samplers.WebsocketMessageSampler
-import com.jmeter.websocket.plugin.samplers.WebsocketMessageSamplerGui
-import com.jmeter.websocket.plugin.samplers.WebsocketSessionSampler
-import com.jmeter.websocket.plugin.samplers.WebsocketSessionSamplerGui
+import com.jmeter.websocket.plugin.elements.samplers.WebsocketMessageSampler
+import com.jmeter.websocket.plugin.elements.samplers.WebsocketMessageSamplerGui
+import com.jmeter.websocket.plugin.elements.samplers.WebsocketSessionSampler
+import com.jmeter.websocket.plugin.elements.samplers.WebsocketSessionSamplerGui
 import spock.lang.Subject
 import spock.lang.Unroll
