+++ src/test/groovy/com/jmeter/websocket/plugin/elements/samplers/modifiers/WebsocketSessionSamplerModifierSpec.groovy
@@ -1,8 +1,8 @@
-package com.jmeter.websocket.plugin.modifiers
+package com.jmeter.websocket.plugin.elements.samplers.modifiers
 
 import com.jmeter.websocket.plugin.JmeterAbstractSpec
-import com.jmeter.websocket.plugin.samplers.WebsocketSessionSampler
-import com.jmeter.websocket.plugin.samplers.WebsocketSessionSamplerGui
+import com.jmeter.websocket.plugin.elements.samplers.WebsocketSessionSampler
+import com.jmeter.websocket.plugin.elements.samplers.WebsocketSessionSamplerGui
 import spock.lang.Subject
 import spock.lang.Unroll
